+++ MSDynamicsDrawerViewController/MSDynamicsDrawerPaneLayout.m
@@ -8,6 +8,7 @@
 
 #import "MSDynamicsDrawerPaneLayout.h"
 #import "MSDynamicsDrawerHelperFunctions.h"
+#import <tgmath.h>
 
 @interface MSDynamicsDrawerPaneLayout ()
 
@@ -40,7 +41,7 @@
 
     CGFloat revealDistance = 0.0;
     if (closedCenterComponent && openCenterComponent) {
-        revealDistance = fabsf(*openCenterComponent - *closedCenterComponent);
+        revealDistance = fabs(*openCenterComponent - *closedCenterComponent);
     }
     return revealDistance;
 }
@@ -200,7 +201,7 @@ static CGFloat const MSRubberBandingCoefficient = .055;
         
         CGSize size = self.paneContainerView.bounds.size;
         CGFloat sizeComponent = *MSSizeComponentForDrawerDirection(&size, direction);
-        CGFloat distancePastBoundedCenter = fabsf(*paneCenterComponent - *relevantBoundingComponent);
+        CGFloat distancePastBoundedCenter = fabs(*paneCenterComponent - *relevantBoundingComponent);
         CGFloat sizeNormalizedElasticityCoefficient = (sizeComponent * MSRubberBandingCoefficient);
         CGFloat elasticOffset = (sizeNormalizedElasticityCoefficient * logf( (distancePastBoundedCenter / sizeNormalizedElasticityCoefficient) + 1.0));
         CGFloat elasticOffsetSign = ((panTranslationComponent > 0.0) ? 1.0 : -1.0);
@@ -227,7 +228,7 @@ static CGFloat const MSRubberBandingCoefficient = .055;
     CGFloat * const centerClosedComponent = MSPointComponentForDrawerDirection(&paneCenterClosed, direction);
     
     if (centerComponent && centerClosedComponent) {
-        return fabsf(*centerComponent - *centerClosedComponent);
+        return fabs(*centerComponent - *centerClosedComponent);
     }
     return 0.0;
 }
